aura-voice-assistant/requirements.txt
speechrecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.11
gtts==2.3.2
playsound==1.3.0
requests==2.31.0
pyyaml==6.0
python-dotenv==1.0.0